requirements.txt
firebase-admin~=6.6.0
pandas~=2.2.3
requests~=2.32.3
python-dotenv~=1.0.1
streamlit~=1.40.2